[math] Default Solve tolerances
* Sort test methods
